requirements.txt
discord.py==1.7.3
aiohttp==3.8.1
aiosqlite==0.17.0
cachetools==4.2.4
python-dotenv==0.19.1
aiolimiter==0.3.0
